Import students, courses, scoreBoard

+++ Project/sha256.cpp
@@ -111,10 +111,10 @@ void sha256_update(SHA256_CTX* ctx, string data, size_t len)
 
 	}
 
-	ctx->bitlen += ctx->datalen * 8;
+	ctx->bitlen += (unsigned long long)ctx->datalen *(unsigned long long) 8;
 
 
-	for (i = 0; i < 8; i++)ctx->data[63 - i] = ctx->bitlen >> (8 * i);
+	for (i = 0; i < 8; i++)ctx->data[63 - i] = (BYTE)ctx->bitlen >> (8 * i);
 
 	sha256_transform(ctx, ctx->data);
